Enum cargo_credential::error::Error

#[non_exhaustive]
pub enum Error {
    UrlNotSupported,
    NotFound,
    OperationNotSupported,
    Other(Box<dyn StdError + Sync + Send>),
    Unknown,
}

Credential provider error type.

UrlNotSupported and NotFound errors both cause Cargo to attempt another provider, if one is available. The other variants are fatal.

Note: Do not add a tuple variant, as it cannot be serialized.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

UrlNotSupported

Registry URL is not supported. This should be used if the provider only works for some registries. Cargo will try another provider, if available

NotFound

Credentials could not be found. Cargo will try another provider, if available

OperationNotSupported

The provider doesnโt support this operation, such as a provider that canโt support โloginโ / โlogoutโ

Other(Box<dyn StdError + Sync + Send>)

The provider failed to perform the operation. Other providers will not be attempted

Unknown

A new variant was added to this enum since Cargo was built